Separate BeamData.h into smaller header files
