+++ source/main/physics/BeamData.h
@@ -281,83 +281,8 @@ static const float flapangles[6] = {0.f, -0.07f, -0.17f, -0.33f, -0.67f, -1.f};
 
 /* basic structures */
 
-/**
-* SIM-CORE; Node.
-*/
-struct node_t
-{
-	Ogre::Vector3 RelPosition; //!< relative to the local physics origin (one origin per truck) (shaky)
-	Ogre::Vector3 AbsPosition; //!< absolute position in the world (shaky)
-	Ogre::Vector3 Velocity;
-	Ogre::Vector3 Forces;
-	Ogre::Real inverted_mass;
-	Ogre::Real mass;
-	Ogre::Vector3 lastNormal;
-	int locked;
-	int iswheel; //!< 0=no, 1, 2=wheel1  3,4=wheel2, etc...
-	int wheelid; //!< Wheel index
-	int masstype; //!< Loaded (by vehicle cargo)? {0/1}
-	int wetstate; //!< {DRY | DRIPPING | WET}
-	int contactless; //!< Bool{0/1}
-	int lockednode;
-	int lockgroup;
-	Ogre::Vector3 lockedPosition; //!< absolute
-	Ogre::Vector3 lockedForces;
-	Ogre::Vector3 lockedVelocity;
-	int contacted; //!< Boolean
-	Ogre::Real friction_coef;
-	Ogre::Real buoyancy;
-	Ogre::Real volume_coef;
-	Ogre::Real surface_coef;
-	Ogre::Vector3 lastdrag;
-	Ogre::Vector3 gravimass;
-	float wettime; //!< Cumulative time this node has been in contact with water. When wet, produces dripping particles.
-	bool isHot; //!< Makes this node emit vapour particles when in contact with water.
-	bool overrideMass;
-	bool disable_particles;
-	bool disable_sparks;
-	Ogre::Vector3 buoyanceForce;
-	int id; //!< Numeric identifier assigned in rig-definition file (if used), or -1 if the node was generated dynamically.
-	int collisionBoundingBoxID;
-	float collRadius;
-	float collTestTimer;
-	Ogre::Vector3 iPosition; //!< initial position, absolute
-	Ogre::Real    iDistance; //!< initial distance from node0 during loading - used to check for loose parts
-	Ogre::Vector3 smoothpos; //!< absolute, per-frame smooth, must be used for visual effects only
-	bool iIsSkin;
-	bool isSkin;
-	bool contacter;
-	int mouseGrabMode;           //!< { 0=Mouse grab, 1=No mouse grab, 2=Mouse grab with force display}
-	int pos;                     //!< This node's index in rig_t::nodes array.
-	Ogre::SceneNode *mSceneNode; //!< visual  
-};
-
-
-/**
-* SIM-CORE; Shock.
-*/
-struct shock_t
-{
-	int beamid;
-	int flags;
-	float lastpos;
-	float springin;
-	float dampin;
-	float sprogin;
-	float dprogin;
-	float springout;
-	float dampout;
-	float sprogout;
-	float dprogout;
-	float sbd_spring;               //!< set beam default for spring
-	float sbd_damp;                 //!< set beam default for damping
-	int trigger_cmdlong;            //!< F-key for trigger injection longbound-check
-	int trigger_cmdshort;           //!< F-key for trigger injection shortbound-check
-	bool trigger_enabled;           //!< general trigger,switch and blocker state
-	float trigger_switch_state;     //!< needed to avoid doubleswitch, bool and timer in one
-	float trigger_boundary_t;       //!< optional value to tune trigger_switch_state autorelease
-	int last_debug_state;           //!< smart debug output
-};
+#include "datatypes/node.h"
+#include "datatypes/shock.h"
 
 struct collcab_rate_t
 {
@@ -367,67 +292,8 @@ struct collcab_rate_t
 	bool calcforward;
 };
 
-/**
-* SIM-CORE; Beam data.
-*/
-struct beam_t
-{
-	node_t *p1;
-	node_t *p2;
-	Beam *p2truck; //!< in case p2 is on another truck
-	bool disabled;
-	Ogre::Real k; //!< tensile spring
-	Ogre::Real d; //!< damping factor
-	Ogre::Real L; //!< length
-	Ogre::Real minmaxposnegstress;
+#include "datatypes/beam.h"
 
-	//! Beam type (unnamed enum) { BEAM_NORMAL=0, BEAM_HYDRO=1, BEAM_VIRTUAL=2, BEAM_MARKED=3, BEAM_INVISIBLE=4, BEAM_INVISIBLE_HYDRO=5 }
-	int type;
-
-	Ogre::Real maxposstress;
-	Ogre::Real maxnegstress;
-	Ogre::Real shortbound;
-	Ogre::Real longbound;
-	Ogre::Real strength;
-	Ogre::Real stress;
-
-	//! Values (unnamed enum) { SHOCK1=1, SHOCK2=2, SUPPORTBEAM=3, ROPE=4 } 
-	int bounded; 
-
-	bool broken;
-	Ogre::Real plastic_coef;
-	Ogre::Real refL;       //!< reference length
-	Ogre::Real Lhydro;     //!< hydro reference len
-	Ogre::Real hydroRatio; //!< hydro rotation ratio
-	int hydroFlags;
-	int animFlags;
-	float animOption;
-	Ogre::Real commandRatioLong;
-	Ogre::Real commandRatioShort;
-	Ogre::Real commandShort; //<! Max. contraction; proportional to orig. length
-	Ogre::Real commandLong;  //<! Max. extension; proportional to orig. length
-	Ogre::Real commandEngineCoupling;
-	Ogre::Real maxtiestress;
-	Ogre::Real diameter;
-	bool commandNeedsEngine;
-	int detacher_group;	//!< Attribute: detacher group number (integer)
-	Ogre::Vector3 lastforce;
-	bool isCentering;
-	int isOnePressMode;
-	bool isForceRestricted;
-	float iStrength; //!< initial strength
-	Ogre::Real default_deform;
-	Ogre::Real default_plastic_coef;
-	int autoMovingMode;
-	bool autoMoveLock;
-	bool pressedCenterMode;
-	float centerLength;
-	float minendmass;
-	float scale;
-	shock_t *shock;
-	Ogre::SceneNode *mSceneNode; //!< visual
-	Ogre::Entity *mEntity; //!< visual
-};
 
 struct soundsource_t
 {
@@ -456,64 +322,7 @@ struct rigidifier_t
 	beam_t *beamc;
 };
 
-/**
-* SIM-CORE; Wheel.
-*/
-struct wheel_t
-{
-	int nbnodes;
-	node_t* nodes[50];
-	/**
-	 * Defines the braking characteristics of a wheel. Wheels are braked by three mechanisms:
-	 * A footbrake, a handbrake/parkingbrake, and directional brakes used for skidsteer steering.
-	 * - 0 = no  footbrake, no  handbrake, no  direction control -- wheel is unbraked
-	 * - 1 = yes footbrake, yes handbrake, no  direction control
-	 * - 2 = yes footbrake, yes handbrake, yes direction control (braked when truck steers to the left)
-	 * - 3 = yes footbrake, yes handbrake, yes direction control (braked when truck steers to the right)
-	 * - 4 = yes footbrake, no  handbrake, no  direction control -- wheel has footbrake only, such as with the front wheels of a normal car
-	 **/
-	int braked;
-	node_t* arm;
-	node_t* near_attach;
-	node_t* refnode0;
-	node_t* refnode1;
-	int propulsed;
-	Ogre::Real radius;
-	Ogre::Real speed;
-	Ogre::Real delta_rotation; //!< Difference in wheel position
-	float rp;
-	float rp1; //<! Networking; triple buffer
-	float rp2; //<! Networking; triple buffer
-	float rp3; //<! Networking; triple buffer
-	float width;
-
-	// for skidmarks
-	Ogre::Vector3 lastContactInner;
-	Ogre::Vector3 lastContactOuter;
-	Ogre::Vector3 lastRotationVec;
-	bool firstLock;
-	float lastSlip;
-	int lastContactType;
-	ground_model_t *lastGroundModel;
-
-	//skidmarks v2
-	bool isSkiding;
-	
-	// for improved collision code
-	int lastEventHandler;
-};
-
-/**
-* SIM-CORE; Visual wheel.
-*/
-struct vwheel_t
-{
-	node_t *p1;
-	node_t *p2;
-	Flexable *fm;
-	Ogre::SceneNode *cnode;
-	bool meshwheel;
-};
+#include "datatypes/wheel.h"
 
 struct hook_t
 {
@@ -701,304 +510,8 @@ struct debugtext_t
 	Ogre::SceneNode *node;
 };
 
-/**
-* SIM-CORE; Represents a vehicle.
-*/
-struct rig_t
-{
-	// TODO: sort these a bit more ...
-	node_t nodes[MAX_NODES];
-	int free_node;
+#include "datatypes/rig.h"
 
-	beam_t beams[MAX_BEAMS];
-	int free_beam;
-
-	contacter_t contacters[MAX_CONTACTERS];
-	int free_contacter;
-
-	rigidifier_t rigidifiers[MAX_RIGIDIFIERS];
-	int free_rigidifier;
-
-	wheel_t wheels[MAX_WHEELS];
-	vwheel_t vwheels[MAX_WHEELS];
-	int free_wheel;
-
-	std::vector <rope_t> ropes;
-	std::vector <ropable_t> ropables;
-	std::vector <tie_t> ties;
-	std::vector <hook_t> hooks;
-
-	wing_t wings[MAX_WINGS];
-	int free_wing;
-		
-	command_t commandkey[MAX_COMMANDS + 10]; // 0 for safety
-	
-	rotator_t rotators[MAX_ROTATORS];
-	int free_rotator;
-
-	std::vector<flare_t> flares;
-	int free_flare;
-
-	prop_t props[MAX_PROPS];
-	prop_t *driverSeat;
-	int free_prop;
-	
-	shock_t shocks[MAX_SHOCKS];
-	int free_shock;
-	int free_active_shock; //!< this has no array associated with it. its just to determine if there are active shocks!
-
-	std::vector < exhaust_t > exhausts;
-
-	cparticle_t cparticles[MAX_CPARTICLES];
-	int free_cparticle;
-	
-	std::vector<debugtext_t>nodes_debug, beams_debug;
-	
-	soundsource_t soundsources[MAX_SOUNDSCRIPTS_PER_TRUCK];
-	int free_soundsource;
-	
-	int pressure_beams[MAX_PRESSURE_BEAMS];
-	int free_pressure_beam;
-
-	AeroEngine *aeroengines[MAX_AEROENGINES];
-	int free_aeroengine;
-
-	Screwprop *screwprops[MAX_SCREWPROPS];
-	int free_screwprop;
-
-	int cabs[MAX_CABS*3];
-	int subisback[MAX_SUBMESHES]; //!< Submesh; {0, 1, 2}
-	int free_cab;
-
-	int hydro[MAX_HYDROS];
-	int free_hydro;
-
-	Ogre::Vector3 texcoords[MAX_TEXCOORDS];
-	int free_texcoord;
-
-	int subtexcoords[MAX_SUBMESHES];
-	int subcabs[MAX_SUBMESHES];
-	int free_sub;
-
-	int collcabs[MAX_CABS];
-	int collcabstype[MAX_CABS];
-	collcab_rate_t inter_collcabrate[MAX_CABS];
-	collcab_rate_t intra_collcabrate[MAX_CABS];
-	int free_collcab;
-
-	int buoycabs[MAX_CABS];
-	int buoycabtypes[MAX_CABS];
-	int free_buoycab;
-
-	Airbrake *airbrakes[MAX_AIRBRAKES];
-	int free_airbrake;
-
-	Skidmark *skidtrails[MAX_WHEELS*2];
-	bool useSkidmarks;
-
-	FlexBody *flexbodies[MAX_FLEXBODIES];
-	int free_flexbody;
-
-	std::vector <VideoCamera *> vidcams;
-
-	std::vector<std::string> description;
-
-	int cameraRail[MAX_CAMERARAIL];
-	int free_camerarail;
-
-	bool hideInChooser;
-
-	char guid[128];
-	int hasfixes;
-	int wingstart;
-	
-	Ogre::String realtruckname;
-	bool loading_finished;
-
-	bool forwardcommands;
-	bool importcommands;
-	bool wheel_contact_requested;
-	bool rescuer;
-	bool disable_default_sounds;
-
-	// Antilockbrake + Tractioncontrol
-	bool slopeBrake;
-	float slopeBrakeFactor;
-	float slopeBrakeAttAngle;
-	float slopeBrakeRelAngle;
-	float previousCrank;
-	float alb_ratio;        //!< Anti-lock brake attribute: Regulating force
-	float alb_minspeed;     //!< Anti-lock brake attribute;
-	int alb_mode;           //!< Anti-lock brake status; Enabled? {1/0}
-	unsigned int alb_pulse; //!< Anti-lock brake attribute;
-	bool alb_pulse_state;   //!< Anti-lock brake status;
-	bool alb_present;       //!< Anti-lock brake attribute: Display the dashboard indicator?
-	bool alb_notoggle;      //!< Anti-lock brake attribute: Disable in-game toggle?
-	float tc_ratio;
-	float tc_wheelslip;
-	float tc_fade;
-	int tc_mode;           //!< Traction control status; Enabled? {1/0}
-	unsigned int tc_pulse; //!< Traction control attribute;
-	bool tc_pulse_state;
-	bool tc_present;       //!< Traction control attribute; Display the dashboard indicator?
-	bool tc_notoggle;      //!< Traction control attribute; Disable in-game toggle?
-	float tcalb_timer;
-	int antilockbrake;
-	int tractioncontrol;
-	float animTimer;
-
-	// Cruise Control
-	bool cc_mode;
-	bool cc_can_brake;
-	float cc_target_rpm;
-	float cc_target_speed;
-	float cc_target_speed_lower_limit;
-
-	// Speed Limiter
-	bool sl_enabled; //!< Speed limiter;
-	float sl_speed_limit; //!< Speed limiter;
-
-	char uniquetruckid[256];
-	int categoryid;
-	int truckversion;
-	int externalcameramode, externalcameranode;
-	std::vector<authorinfo_t> authors;
-	float fadeDist;
-	float collrange;
-	int masscount; //!< Number of nodes loaded with l option
-	bool disable_smoke;
-	int smokeId;  //!< Old-format exhaust (one per vehicle) emitter node
-	int smokeRef; //!< Old-format exhaust (one per vehicle) backwards direction node
-	char truckname[256];
-	bool networking;
-	int editorId;
-	bool beambreakdebug, beamdeformdebug, triggerdebug;
-	CmdKeyInertia *rotaInertia;
-	CmdKeyInertia *hydroInertia;
-	CmdKeyInertia *cmdInertia;
-	bool enable_wheel2; //!< If false, wheels2 are downgraded to wheels1 (needed for multiplayer)
-	float truckmass;
-	float loadmass;
-	char texname[1024]; //!< Material name
-	int trucknum;
-	Skin *usedSkin;
-	Buoyance *buoyance;
-
-	int driveable;
-	BeamEngine *engine;
-	int hascommands;
-	int hashelp;
-	char helpmat[256];
-	int cinecameranodepos[MAX_CAMERAS]; //!< Cine-camera node indexes
-	int freecinecamera; //!< Number of cine-cameras (lowest free index)
-	int flaresMode;
-	Ogre::Light *cablight;
-	Ogre::SceneNode *cablightNode;
-	std::vector<Ogre::Entity*> deletion_Entities; //!< For unloading vehicle; filled at spawn.
-	std::vector<Ogre::MovableObject *> deletion_Objects; //!< For unloading vehicle; filled at spawn.
-	std::vector<Ogre::SceneNode*> deletion_sceneNodes; //!< For unloading vehicle; filled at spawn.
-	int netCustomLightArray[4];
-	unsigned char netCustomLightArray_counter;
-	MaterialFunctionMapper *materialFunctionMapper;
-	bool driversseatfound;
-	bool ispolice;
-	int state;
-	bool collisionRelevant;
-	bool heathaze;
-	Autopilot *autopilot;
-	HeightFinder *hfinder;
-	Airfoil *fuseAirfoil;
-	node_t *fuseFront;
-	node_t *fuseBack;
-	float fuseWidth;
-	float brakeforce;
-	float hbrakeforce;
-	//! Dbg. overlay type { NODES: 1-Numbers, 4-Mass, 5-Locked | BEAMS: 2-Numbers, 6-Compression, 7-Broken, 8-Stress, 9-Strength, 10-Hydros, 11-Commands, OTHER: 3-N&B numbers, 12-14 unknown }
-	int debugVisuals;
-	
-	Ogre::String speedomat, tachomat;
-	float speedoMax;
-	bool useMaxRPMforGUI;
-	float minimass;
-	bool cparticle_enabled;
-	bool advanced_drag;
-	float advanced_node_drag;
-	float advanced_total_drag;
-
-	Axle *axles[MAX_WHEELS/2];
-	int free_axle;
-
-	int free_fixes;
-	int propwheelcount;
-	int free_commands;
-	int fileformatversion;
-
-	std::vector<Ogre::String> sectionconfigs;
-
-	Ogre::Vector3 origin;
-	Ogre::SceneNode *beamsRoot;
-	//! Stores all the SlideNodes available on this truck
-	std::vector< SlideNode > mSlideNodes;
-
-	int proped_wheels; //!< Number of propelled wheels.
-	int braked_wheels; //!< Number of braked wheels.
-	
-	int proppairs[MAX_WHEELS]; //!< For inter-differential locking
-
-	//! try to connect slide-nodes directly after spawning
-	bool slideNodesConnectInstantly;
-
-	//! Stores all the available RailGroups for this truck
-	std::vector< RailGroup* > mRailGroups;
-
-	Ogre::Camera *mCamera;
-	int freecamera;
-	int cameranodepos[MAX_CAMERAS];
-	int cameranodedir[MAX_CAMERAS];
-	int cameranoderoll[MAX_CAMERAS];
-	bool revroll[MAX_CAMERAS];
-	bool shadowOptimizations;
-	int hasEmissivePass;
-	FlexObj *cabMesh;
-	Ogre::SceneNode *cabNode;
-	Ogre::AxisAlignedBox boundingBox; //!< standard bounding box (surrounds all nodes of a truck)
-	Ogre::AxisAlignedBox predictedBoundingBox;
-	std::vector<Ogre::AxisAlignedBox> collisionBoundingBoxes; //!< smart bounding boxes, used for determining the state of a truck (every box surrounds only a subset of nodes)
-	std::vector<Ogre::AxisAlignedBox> predictedCollisionBoundingBoxes;
-	bool freePositioned;
-	int lowestnode; //!< never updated after truck init!?!
-
-	float default_spring; //!< TODO: REMOVE! (parser context only)
-	float default_spring_scale; //!< TODO: REMOVE! (parser context only)
-	float default_damp;
-	float default_damp_scale;
-	float default_deform;
-	float default_deform_scale;
-	float default_break;
-	float default_break_scale;
-
-	float default_beam_diameter;
-	float default_plastic_coef;
-	float skeleton_beam_diameter;
-
-	char default_beam_material[256];
-	float default_node_friction; //!< TODO: REMOVE! (parser context only)
-	float default_node_volume; //!< TODO: REMOVE! (parser context only)
-	float default_node_surface; //!< TODO: REMOVE! (parser context only)
-	float default_node_loadweight; //!< TODO: REMOVE! (parser context only)
-	char default_node_options[50]; //!< TODO: REMOVE! (parser context only)
-
-	float posnode_spawn_height;
-
-	MaterialReplacer *materialReplacer;
-	Ogre::String subMeshGroundModelName;
-
-	float odometerTotal;
-	float odometerUser;
-
-	std::vector<std::pair<Ogre::String, bool> > dashBoardLayouts;
-	Ogre::String beamHash; //!< Unused
-};
 
 // some non-beam structs
